PR Feedback: Pulling nodeUniqueConstraintTableDef out into a new file, constraint_table_def.go
