+++ server/ast/alter_table.go
@@ -107,45 +107,3 @@ func nodeAlterTableAddConstraint(
 			"definition type %T", node)
 	}
 }
-
-// nodeUniqueConstraintTableDef converts a tree.UniqueConstraintTableDef instance
-// into a vitess.DDL instance that can be executed by GMS. |tableName| identifies
-// the table being altered, and |ifExists| indicates whether the IF EXISTS clause
-// was specified.
-func nodeUniqueConstraintTableDef(
-	node *tree.UniqueConstraintTableDef,
-	tableName vitess.TableName,
-	ifExists bool) (*vitess.DDL, error) {
-
-	if len(node.IndexParams.StorageParams) > 0 {
-		return nil, fmt.Errorf("STORAGE parameters not yet supported for indexes")
-	}
-
-	if node.IndexParams.Tablespace != "" {
-		return nil, fmt.Errorf("TABLESPACE is not yet supported")
-	}
-
-	if node.NullsNotDistinct {
-		return nil, fmt.Errorf("NULLS NOT DISTINCT is not yet supported")
-	}
-
-	columns, err := nodeIndexElemList(node.Columns)
-	if err != nil {
-		return nil, err
-	}
-
-	if node.PrimaryKey {
-		return &vitess.DDL{
-			Action:   "alter",
-			Table:    tableName,
-			IfExists: ifExists,
-			IndexSpec: &vitess.IndexSpec{
-				Action:  "create",
-				Type:    "primary",
-				Columns: columns,
-			},
-		}, nil
-	} else {
-		return nil, fmt.Errorf("Only PRIMARY KEY constraints are supported currently")
-	}
-}
